FE-27 Arreglos del merge con el widget Text
